scheduler: Be able to handle no running tasks
Right now, if we ever get into this state we'll never leave, since
the only things that can run (interrupts) never schedule new tasks.

The idea is that as mechanisms for tasks to block are implemented,
the scheduler will check for tasks to become unblocked in each iteration
of the halt loop.