scheduler: Add data structures for tracking sleeping tasks
Wtih this scheme, we only check the front task (so it's fast to check)
and we aren't storing a full timestamp per sleeping task (just a delta).
Waking a task will be slow if there's a long list of sleepers behind it,
but we need the ability to insert into the middle of the array when adding a sleeper.

Adding is not implemented, and will be hard and slow.
Checking and removing is implemented.