Refactor points where we suspend a task to either re-enqueue it to run or not
Sometimes we don't want the task to run again, eg. when we're putting it to sleep.
We now only have two entry points to suspend a running task:
TaskYield - Suspend task but keep it runnable (re-enqueue it)
SchedSleepTask - Put task to sleep for a given interval, and don't re-enqueue it