graphics: Disable vblank interrupt entirely if we have nothing to do
Every time vblank handler completes with time to spare, it disables the interrupt.
Every time we add to the graphics queue, it enables the interrupt.

This can result in a spurious run if it clears the queue and disables itself
after the items are written but before the enable line runs, but this isn't a problem.