tasks: Allow setting initial ROM bank
Without being able to set a ROM bank, tasks must resort to shims
that load the correct bank then jump to it at the start of the task.
This is annoying boilerplate and not nessecary.

Instead, we just set an initial bank during task creation.
We still require the task set its own initial RAM bank but this is not as difficult.