task clock: Display uptime in top-left of screen with sprites
It's a bit jerky at the moment. Curiously it appears to be way smoother when
we busy loop (never call sleep) than even if we sleep(1) (which means we're pretty much
always runnable as soon as we check). Are we getting pushed to the back of the queue or something?
shouldn't matter...