unit test gen: Emit helpful debug messages
To prepare for more automated testing, and to assist in general,
we add BGB debug statements detailing:
	Success, or
	Failure, and the mismatched value, expected and actual values.

Note we weren't able to do this in the most straightforward manner of hardcoding
every log line for the Memory checks, which can sometimes be 100s of items long.
Having a full debug string for every index made the ROM too big to assemble.

Instead, we track the current index in DE and put the expected value in B before
comparing it, so we can emit those values in the debug string programatically.