CheckNextWake: Fix a typo that completely broke things when multiple tasks sleeping
The RepointStruct was the wrong way around, so it was copying everything
up 1 byte instead of 3 bytes (dec instead of inc when setting initial positions).
Not only does this fill the queue with garbage, but it means we're relying on
uninitialized memory to make us stop looping, which can eventually cause
all kinds of weirdness. Pretty sure that in at least one test I corrupted my stack
and returned into a random part of memory.