debug: Add a common header to every debug message
with current file and clocks since startup.

I debated using Uptime instead of BGB's internal clock,
but it won't always be perfectly accurate, particularly in a debug message
when I might be debugging issues with the Uptime clock itself,
or in a race after TimerCounter has wrapped but before the timer interrupt
has fired/finished.

I also attempted to include line number but this does not work currently:
	https://github.com/rednex/rgbds/issues/206