scheduler: Fix a bug where T_SchedSleepTask would sleep for bogus lengths
The means by which we were saving and retrieving DE were vulnerable to a race
with an interrupt firing causing them to be overwritten:
	1. We set CoreStack-1,CoreStack-2
	2. In TaskSave, we load CoreStack into SP
	3. We get an interrupt, writing current PC to CoreStack-1 and CoreStack-2
	4. We load CoreStack-1 and CoreStack-2 into DE

Solution: Massively simplify and change contract so clobbering all is ok,
then simply perform the EnqueueSleepTask BEFORE the TaskSave, so preserving
DE is no longer an issue.
While there, removed non-T_ version of SchedSleepTask, since its use was
rather nebulous, and it had no users.