Instead of explicitly setting aligned ram area, use new ALIGN feature, and put it in WRAM0
ALIGN[8] will tell assembler (actually, the linker) that it needs to put the start
of this section at an address aligned to 8 address bits (ie. 256 bytes).

This lets us also put it in WRAM0 where it was always meant to be.
It was only in WRAMX to ensure it had a good alignment area off on its own.