Initial implementation of waiters
Waiter struct maintains a count and task id of the lowest task,
so that in the case of 1 task it's very quick. Otherwise you have to scan but
at least you can skip the first few tasks.

Also saves a unique value, a 'determinant' in the task_waiter field.
It's the presence of this value that indicates a task is waiting on a given waiter.

When a task goes away, it need only invalidate its tasks entry, and it won't be woken.

Right now the determinant is a combination of the bank and most bits of the address,
packing in some weird ways and making use of the fact that addresses will only ever be in
WRAM, SRAM or HRAM.

However, in hindsight it occurs that a task that's waiting on a value in a given bank
must nessecarily have that bank loaded. In this way we can simply use the addr as the determinant,
and check the bank value against the current one.