Create an int-safe wrapper of waiter
Waiters are not interrupt-safe, but some use-cases require waking from an interrupt,
eg. graphics queues, joypad queue

We implement this by adding a single byte flag with the following meanings:
0: resting state. safe to wake.
1: a wait is in the process of being added. if you're trying to wake, set to 2 and do nothing
2: a wake attempted to occur while in state 1

When the task trying to wait sees a 2 after adding, it immediately wakes all tasks (including itself).

Also included is a version that keeps 1 set while performing the "should we wait" check condition.
This is needed to prevent a race like:
1. check is made, condition indicates we should wait
2. condition changes, we shouldn't wait, and waiters are woken
3. we wait